pyproject.toml
[project]
name = "embodied-nav"
version = "0.1.0"
description = "LLM-powered embodied navigation system"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "opencv-python>=4.8.0",
    "httpx>=0.25.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
embodied-nav = "embodied_nav.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/embodied_nav"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
